// src/models/SupplierModel.ts
/** @format */

export interface SupplierModel {
  index: number;
  name: string;
  slug: string;
  product: string;
  categories: string[];
  price: number;
  contact: string;
  isTaking: number;
  photoUrl: string;
  createdAt: string;
  updatedAt: string;
  email: string;
  active: string;
  _id: string;
}